# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Taskify-API",
        "description": "Description de votre API",
        "contact": {
            "name": "Mohamed EL MORJANI",
            "email": "elmorjanimohamed9@email.com"
        },
        "license": {
            "name": "Licence MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout the current user",
                "description": "Logout the authenticated user and revoke the current access token",
                "operationId": "cfadf87dd76235e9475d9f843f55ba77",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}